Fix invoke run log module resolution
